Made lots of changes

+++ Backend/routes/houses.js
@@ -10,8 +10,12 @@ const {
   updateHouseInDb,
   toggleBayAssignment,
 } = require("../services/houseServices");
+const { getLogsByHouseId } = require("../services/logServices");
 
-const {requireLoggin, requirePermission} = require('../middleware/authMiddleware')
+const {
+  requireLoggin,
+  requirePermission,
+} = require("../middleware/authMiddleware");
 
 // GET endpoint to retrieve all houses
 router.get("/", requireLoggin, async (req, res) => {
@@ -43,7 +47,6 @@ router.get("/inbay", requireLoggin, async (req, res) => {
 // GET endpoint to retrieve the house that is in a specific bay
 router.get("/inbay/:bayid", requireLoggin, async (req, res) => {
   const { bayid } = req.params;
-  console.log("bayid", bayid);
   try {
     const house = await getHouseInBay(bayid);
     res.json({ result: house });
@@ -53,42 +56,57 @@ router.get("/inbay/:bayid", requireLoggin, async (req, res) => {
 });
 
 // GET endpoint to retrieve a specific house
-router.get("/:houseid", requireLoggin,  async (req, res) => {
+router.get("/:houseid", requireLoggin, async (req, res) => {
   const { houseid } = req.params;
   try {
-    const house = await getHouseFromDb(houseid); // Function to fetch a specific house
+    const houseLogs = await getLogsByHouseId(houseid);
+    const house = (await getHouseFromDb(houseid))[0];
+    house.logs = houseLogs;
     if (!house) {
       res.status(404).send("House not found");
     } else {
-      res.json({ result: house[0] });
+      res.json({ result: house });
     }
   } catch (error) {
+    console.log("error", error);
     res.status(500).send("Server error");
   }
 });
 
 // POST endpoint to add a new house
-router.post("/", requireLoggin, requirePermission('admin'), async (req, res) => {
-  try {
-    const newHouse = await addHouseToDb(req.body); // Function to add a new house
-    res.status(201).json({ result: newHouse });
-  } catch (error) {
-    console.log("error", error);
-    res.status(500).send("Server error", error);
+router.post(
+  "/",
+  requireLoggin,
+  requirePermission("admin"),
+  async (req, res) => {
+    try {
+      const newHouse = await addHouseToDb(req.body);
+      const houseLogs = await getLogsByHouseId(newHouse._id);
+      newHouse.logs = houseLogs;
+      res.status(201).json({ result: newHouse });
+    } catch (error) {
+      console.log("error", error);
+      res.status(500).send("Server error", error);
+    }
   }
-});
+);
 
 // DELETE endpoint to remove a house
-router.delete("/:houseid",requireLoggin, requirePermission('admin'),  async (req, res) => {
-  const { houseid } = req.params;
-  try {
-    const houseDeleted = await deleteHouseFromDb(houseid); // Function to delete a house
-    res.status(200).json({ result: { houseDeleted } });
-  } catch (error) {
-    console.log("error", error);
-    res.status(500).send("Server error", error);
+router.delete(
+  "/:houseid",
+  requireLoggin,
+  requirePermission("admin"),
+  async (req, res) => {
+    const { houseid } = req.params;
+    try {
+      const houseDeleted = await deleteHouseFromDb(houseid); // Function to delete a house
+      res.status(200).json({ result: { houseDeleted } });
+    } catch (error) {
+      console.log("error", error);
+      res.status(500).send("Server error", error);
+    }
   }
-});
+);
 
 // PUT endpoint to update house information
 router.put("/:houseid", requireLoggin, async (req, res) => {
@@ -96,6 +114,8 @@ router.put("/:houseid", requireLoggin, async (req, res) => {
   try {
     const updatedHouse = await updateHouseInDb(houseid, req.body); // Function to update house details
     if (updatedHouse) {
+      const houseLogs = await getLogsByHouseId(updatedHouse._id);
+      updatedHouse.logs = houseLogs;
       res.status(200).json({ result: updatedHouse });
     } else {
       res.status(404).send("House not found");
@@ -111,6 +131,8 @@ router.patch("/:houseid/:bayid", requireLoggin, async (req, res) => {
   try {
     const updatedHouse = await toggleBayAssignment(houseid, bayid); // Function to attach/detach a bay
     if (updatedHouse) {
+      const houseLogs = await getLogsByHouseId(updatedHouse._id);
+      updatedHouse.logs = houseLogs;
       res.status(200).json({ result: updatedHouse });
     } else {
       res.status(404).send("House not found");
